All Schedule screen created - Added new basic elements in data, domain layers. The schedule screen interface has also been created

+++ core/navigation/src/main/java/com/dashkevich/navigation/Navigation.kt
@@ -9,6 +9,6 @@ fun NavController.navigateToMain() {
     }
 }
 
-fun NavController.navigateToSchedule(leagueId: Int) {
+fun NavController.navigateToSchedule(leagueId: Long) {
     navigate("$SCHEDULE/$leagueId")
 }
