+++ features/main/src/main/java/com/dashkevich/main/components/LeaguesCard.kt
@@ -22,14 +22,14 @@ private val fourDp = 4.dp
 private const val two = 2
 
 @Composable
-fun LeaguesCard(modifier: Modifier = Modifier, league: League, onLeagueClick: (Int) -> Unit) {
+fun LeaguesCard(modifier: Modifier = Modifier, league: League, onLeagueClick: (Long) -> Unit) {
     Card(
         modifier = modifier
             .padding(horizontal = 14.dp)
             .fillMaxWidth()
             .height(140.dp)
             .clickable {
-                onLeagueClick(league.leagueId)
+                onLeagueClick(league.leagueId.toLong())
             },
         colors = CardDefaults.cardColors(containerColor = CustomTheme.colors.primary),
         shape = RoundedCornerShape(CustomTheme.shapes.default),
